+++ src/main/java/kr/co/lupintech/controller/AdminController.java
@@ -1,19 +1,21 @@
 package kr.co.lupintech.controller;
 
+import kr.co.lupintech.dto.leave.LeaveRequest;
 import kr.co.lupintech.dto.leave.LeaveResponse;
-import kr.co.lupintech.dto.manager.ManagerRequest;
 import kr.co.lupintech.core.annotation.MyErrorLog;
 import kr.co.lupintech.core.annotation.MyLog;
 import kr.co.lupintech.dto.PageDTO;
-import kr.co.lupintech.dto.manage.ManageResponse;
+import kr.co.lupintech.dto.user.UserRequest;
+import kr.co.lupintech.dto.user.UserResponse;
 import kr.co.lupintech.model.user.User;
 import kr.co.lupintech.service.LeaveService;
-import kr.co.lupintech.service.ManageService;
+import kr.co.lupintech.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 import kr.co.lupintech.dto.ResponseDTO;
 import kr.co.lupintech.model.leave.enums.LeaveStatus;
@@ -24,22 +26,39 @@ import java.util.List;
 @RequiredArgsConstructor
 @RestController
 public class AdminController {
-    private final ManageService manageService;
     private final LeaveService leaveService;
+    private final UserService userService;
 
-    // 유저의 연차 일수 수정
     @PostMapping("/admin/annual/{id}")
-    public ResponseEntity<?> annualUpdate(@PathVariable Long id, @RequestBody @Valid ManagerRequest.AnnualInDTO annualInDTO, Error error) {
-        manageService.연차수정(id, annualInDTO);
+    public ResponseEntity<?> annualUpdate(@PathVariable Long id, @RequestBody @Valid UserRequest.AnnualInDTO annualInDTO) {
+        userService.연차수정(id, annualInDTO);
         ResponseDTO<?>responseDTO = new ResponseDTO<>();
         return ResponseEntity.ok(responseDTO);
     }
 
-    // role까지 변경가능, master로 접근해야, role변경이 가능하다.
-    @PostMapping("/master/{id}")
-    public ResponseEntity<?> roleUpdate(@PathVariable Long id,@RequestBody ManagerRequest.MasterInDTO masterInDTO){
-        manageService.권한수정(id, masterInDTO);
+    @MyLog
+    @MyErrorLog
+    @GetMapping("/admin") // /admin?query=김&page=0&size=10
+    public ResponseEntity<?> search(@RequestParam(defaultValue = "") String query,
+                                    @RequestParam(defaultValue = "0") int page,
+                                    @RequestParam(defaultValue = "10") int size) {
+        Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.ASC, "id"));
+        PageDTO<UserResponse.UserOutDTO, User> pageDTO = userService.사원검색(query, pageable);
+        ResponseDTO<PageDTO<UserResponse.UserOutDTO, User>> responseDTO = new ResponseDTO<>(pageDTO);
+        return ResponseEntity.ok(responseDTO);
+    }
+
+    @PostMapping("/admin/resign/{id}")
+    public ResponseEntity<?> resign(@PathVariable Long id){
+        userService.퇴사(id);
         ResponseDTO<?> responseDTO = new ResponseDTO<>();
+        return ResponseEntity.ok(responseDTO);
+    }
+
+    @PostMapping("/admin/approve")
+    public ResponseEntity<?> decide(@RequestBody @Valid LeaveRequest.DecideInDTO decideInDTO, Errors errors){
+        LeaveResponse.DecideOutDTO decideOutDTO = leaveService.연차당직결정하기(decideInDTO);
+        ResponseDTO<?> responseDTO = new ResponseDTO<>(decideOutDTO);
         return ResponseEntity.ok(responseDTO);
     }
 
@@ -49,18 +68,4 @@ public class AdminController {
         ResponseDTO<List<LeaveResponse.InfoOutDTO>> responseDTO = new ResponseDTO<>(waitingLeaves);
         return ResponseEntity.ok().body(responseDTO);
     }
-
-    // 사원 목록 조회 + 검색
-    @MyLog
-    @MyErrorLog
-    @GetMapping("/admin") // /admin?query=김&page=0&size=10
-    public ResponseEntity<?> search(@RequestParam(defaultValue = "") String query,
-                                    @RequestParam(defaultValue = "0") int page,
-                                    @RequestParam(defaultValue = "10") int size) {
-        Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.ASC, "id"));
-        PageDTO<ManageResponse.UserOutDTO, User> pageDTO = manageService.사원검색(query, pageable);
-        ResponseDTO<PageDTO<ManageResponse.UserOutDTO, User>> responseDTO = new ResponseDTO<>(pageDTO);
-        return ResponseEntity.ok(responseDTO);
-    }
-
 }
