+++ src/main/java/kr/co/lupintech/core/dummy/DummyEntity.java
@@ -52,37 +52,6 @@ public class DummyEntity {
                 .build();
     }
 
-    public User newMockStateUser(Long id, String username, String email, Integer remainDays,boolean status){
-        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-        return User.builder()
-                .id(id)
-                .username(username)
-                .password(passwordEncoder.encode("1234"))
-                .email(email)
-                .role(UserRole.ROLE_USER)
-                .status(status)
-                .hireDate(LocalDate.now().minusYears(1).minusWeeks(1)) // 입사 1년차라 가정
-                .remainDays(remainDays)
-                .createdAt(LocalDateTime.now())
-                .build();
-    }
-
-
-    public User newMockUserRole(Long id, String username, String email, Integer remainDays,UserRole userRole){
-        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-        return User.builder()
-                .id(id)
-                .username(username)
-                .password(passwordEncoder.encode("1234"))
-                .email(email)
-                .role(userRole)
-                .status(true)
-                .hireDate(LocalDate.now().minusYears(1).minusWeeks(1)) // 입사 1년차라 가정
-                .remainDays(remainDays)
-                .createdAt(LocalDateTime.now())
-                .build();
-    }
-
     public Leave newMockLeave(Long id, User user, LeaveType type, LocalDate startDate, LocalDate endDate, Integer usingDays){
         return Leave.builder()
                 .id(id)
@@ -104,16 +73,4 @@ public class DummyEntity {
                 .createdAt(LocalDateTime.now())
                 .build();
     }
-
-    public ManagerRequest.ManageUserListDTO newMockChartUser(Long userId, UserRole role, String username, LocalDate hireDate, Integer remainDays, String profile){
-        return ManagerRequest.ManageUserListDTO.builder()
-                .userId(userId)
-                .role(role)
-                .username(username)
-                .hireDate(hireDate)
-                .remainDays(remainDays)
-                .profile(profile)
-                .build();
-    }
-
 }
