+++ src/main/java/kr/co/lupintech/dto/user/UserResponse.java
@@ -72,4 +72,23 @@ public class UserResponse {
             this.profile = user.getProfile();
         }
     }
+
+    @Setter @Getter
+    public static class UserOutDTO {
+        private Long id; // 유저의 아이디
+        private String username;
+        private UserRole role;
+        private LocalDate hireDate;
+        private Integer remainDays;
+        private String profile;
+
+        public UserOutDTO(User user) {
+            this.id = user.getId();
+            this.username = user.getUsername();
+            this.role = user.getRole();
+            this.hireDate = user.getHireDate();
+            this.remainDays = user.getRemainDays();
+            this.profile = user.getProfile();
+        }
+    }
 }
