+++ src/main/java/kr/co/lupintech/service/UserService.java
@@ -1,11 +1,13 @@
 package kr.co.lupintech.service;
 
+import kr.co.lupintech.core.annotation.MyErrorLog;
 import kr.co.lupintech.core.annotation.MyLog;
 import kr.co.lupintech.core.auth.jwt.MyJwtProvider;
 import kr.co.lupintech.core.auth.session.MyUserDetails;
 import kr.co.lupintech.core.exception.Exception400;
 import kr.co.lupintech.core.exception.Exception401;
 import kr.co.lupintech.core.exception.Exception500;
+import kr.co.lupintech.dto.PageDTO;
 import kr.co.lupintech.dto.user.UserRequest;
 import kr.co.lupintech.dto.user.UserResponse;
 import kr.co.lupintech.model.token.RefreshTokenEntity;
@@ -15,6 +17,8 @@ import kr.co.lupintech.model.user.UserRepository;
 import kr.co.lupintech.service.S3Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.util.Pair;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -24,7 +28,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
@@ -32,10 +38,8 @@ import java.util.Optional;
 public class UserService {
     private final AuthenticationManager authenticationManager;
     private final UserRepository userRepository;
-
     private final TokenRepository tokenRepository;
     private final BCryptPasswordEncoder passwordEncoder;
-
     private final S3Service s3Service;
 
 
@@ -104,7 +108,6 @@ public class UserService {
         return loginOutDTO;
     }
 
-
     @Transactional
     public UserResponse.ModifiedOutDTO 개인정보수정(UserRequest.ModifiedInDTO modifiedInDTO, MultipartFile profile, Long id) {
         // 1. 아이디로 회원 조회
@@ -160,11 +163,46 @@ public class UserService {
         return new UserResponse.ModifiedOutDTO(user, isPasswordReset, isProfileReset);
     }
 
+    @MyLog
+    @MyErrorLog
+    @Transactional
+    public void 연차수정(Long id, UserRequest.AnnualInDTO annualInDTO) {
+        User userPS = userRepository.findByStatusAndId(true, id)
+                .orElseThrow(()->new Exception400("id", "해당 유저가 존재하지 않습니다."));
+        // 정보 수정
+        userPS.setRemainDays(annualInDTO.getRemainDays());
+    }
+
+    @MyLog
+    @MyErrorLog
+    @Transactional
+    public PageDTO<UserResponse.UserOutDTO, User> 사원검색(String query, Pageable pageable) {
+        Page<User> userPG = query.isBlank() ? // 검색할 query가 없으면 전체 목록 조회(퇴사한 회원은 조회x)
+                userRepository.findAllByStatus(true, pageable) : userRepository.findAllByQuery(query, pageable);
+
+        List<UserResponse.UserOutDTO> content = userPG.getContent().stream()
+                .map(user -> new UserResponse.UserOutDTO(user))
+                .collect(Collectors.toList());
+        return new PageDTO<>(content, userPG);
+    }
+
+    @MyLog
+    @MyErrorLog
     @Transactional
     public void 퇴사(Long id) {
-        User userPS = userRepository.findById(id).orElseThrow(
-                () -> new Exception500("로그인 된 유저가 DB에 존재하지 않음")
+        User userPS = userRepository.findByStatusAndId(true, id).orElseThrow(
+                () -> new Exception500("해당 유저가 존재하지 않습니다.")
         );
         userPS.resign();
+    }
+
+    @MyLog
+    @MyErrorLog
+    @Transactional
+    public void 권한수정(Long id, UserRequest.MasterInDTO masterInDTO) {
+        User userPS = userRepository.findByStatusAndId(true, id)
+                .orElseThrow(()->new Exception400("id", "해당 유저가 존재하지 않습니다."));
+        // 정보 수정
+        userPS.setRole(masterInDTO.getRole());
     }
 }
